Move ov::intel_npu::Platform namespace to public properties.hpp
